perf(tests): move two constant sets out of React modules into shared
`model-version.schema.ts` imported MIN/MAX_DONATION_GOAL - two numbers - from
`~/components/Model/ModelVersions/model-version.utils`, and `training.schema.ts`
imported `blockedCustomModels` from `~/components/Training/Form/TrainingCommon`.
Both utils modules are React hooks that reach the session provider and the buzz
purchase tree, so those two edges put ~560 and ~590 modules respectively into
every server graph downstream of model.schema and training.schema.

model-version.schema 605 -> 45 modules, training.schema 607 -> 17.
