+++ src/server/schema/model-version.schema.ts
@@ -5,7 +5,7 @@ import * as z from 'zod';
 import {
   MAX_DONATION_GOAL,
   MIN_DONATION_GOAL,
-} from '~/components/Model/ModelVersions/model-version.utils';
+} from '~/shared/constants/donation-goal.constants';
 import type { BaseModel } from '~/shared/constants/basemodel.constants';
 import { constants } from '~/server/common/constants';
 import { infiniteQuerySchema } from '~/server/schema/base.schema';
